requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorlog==6.9.0
datasets==3.5.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt
python-docx>=0.8.11
dotenv==0.9.9
drf-spectacular==0.28.0
dspy==2.6.16
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
google-cloud-storage==2.16.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json_repair==0.40.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
litellm==1.63.7
lxml==5.3.1
magicattr==0.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.2.0
multiprocess==0.70.16
numpy==2.2.4
openai==1.61.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.1
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0

google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0